+++ lib/include/perceptron.h
@@ -10,13 +10,13 @@
 class Perceptron {
  public:
     Perceptron() = default;
-    Perceptron(const std::vector<int>& aLayers,
+    Perceptron(const std::vector<size_t>& aLayers,
         Neuron::ActivationFunction aFunction =
             Neuron::ActivationFunction::SIGMOID);
 
-    bool initializeNetwork(const std::vector<int>& aLayers,
-        Neuron::ActivationFunction aFunction =
-            Neuron::ActivationFunction::SIGMOID);
+    bool initializeNetwork(const std::vector<size_t> &aLayers,
+                           Neuron::ActivationFunction aFunction =
+                           Neuron::ActivationFunction::SIGMOID);
     bool isConfigured() const;
 
     // NOLINTNEXTLINE(build/include_what_you_use)
